+++ zstandard/backend_cffi.py
@@ -778,6 +778,12 @@ class ZstdCompressionWriter(object):
 
         return False
 
+    def __iter__(self):
+        raise io.UnsupportedOperation()
+
+    def __next__(self):
+        raise io.UnsupportedOperation()
+
     def memory_size(self):
         return lib.ZSTD_sizeof_CCtx(self._compressor._cctx)
 
@@ -3463,6 +3469,12 @@ class ZstdDecompressionWriter(object):
         self.close()
 
         return False
+
+    def __iter__(self):
+        raise io.UnsupportedOperation()
+
+    def __next__(self):
+        raise io.UnsupportedOperation()
 
     def memory_size(self):
         return lib.ZSTD_sizeof_DCtx(self._decompressor._dctx)